Moved uppmaboss server test to test_server.py
